refactor: loading bar & dom root 파일 위치 수정
