+++ src/app.js
@@ -1,6 +1,5 @@
 import { initStore } from './store/index.js';
-import { initRender } from './render.js';
-
+import { initRender } from './template/index.js';
 (async () => await initStore())()
   .then(() => initRender());
